add new preprocessor step but calls still broken
